+++ src/features/commit/service.ts
@@ -1,5 +1,22 @@
-import { octokit } from "./common.js";
-import { ListCommitsArgs, GetCommitArgs } from "../handlers/commit.js";
+import { octokit } from "../../shared/service.js";
+
+export interface ListCommitsArgs {
+  owner: string;
+  repo: string;
+  sha?: string;
+  path?: string;
+  author?: string;
+  since?: string;
+  until?: string;
+  per_page?: number;
+  page?: number;
+}
+
+export interface GetCommitArgs {
+  owner: string;
+  repo: string;
+  ref: string;
+}
 
 export const listCommits = async (params: ListCommitsArgs) => {
   try {
