+++ src/features/file/handler.ts
@@ -1,24 +1,8 @@
-import { GetFileContentParamsSchema, CreateOrUpdateFileParamsSchema } from "../schemas/index.js";
-import { getFileContent, createOrUpdateFile } from "../services/index.js";
+import { GetFileContentParamsSchema, CreateOrUpdateFileParamsSchema } from "./schema.js";
+import { getFileContent, createOrUpdateFile } from "./service.js";
+import { GetFileContentArgs, CreateOrUpdateFileArgs } from "./service.js";
 
 export type HandlerFunction = (args: unknown) => Promise<any>;
-
-export interface GetFileContentArgs {
-  owner: string;
-  repo: string;
-  path: string;
-  ref?: string;
-}
-
-export interface CreateOrUpdateFileArgs {
-  owner: string;
-  repo: string;
-  path: string;
-  message: string;
-  content: string;
-  branch?: string;
-  sha?: string;
-}
 
 export const fileHandlers: Record<string, HandlerFunction> = {
   get_file_content: async (args: unknown) => {
