+++ src/features/pullRequest/handler.ts
@@ -1,35 +1,8 @@
-import { CreatePullRequestParamsSchema, ListPullRequestsParamsSchema, GetPullRequestParamsSchema } from "../schemas/index.js";
-import { createPullRequest, listPullRequests, getPullRequest } from "../services/index.js";
+import { CreatePullRequestParamsSchema, ListPullRequestsParamsSchema, GetPullRequestParamsSchema } from "./schema.js";
+import { createPullRequest, listPullRequests, getPullRequest } from "./service.js";
+import { CreatePullRequestArgs, ListPullRequestsArgs, GetPullRequestArgs } from "./service.js";
 
 export type HandlerFunction = (args: unknown) => Promise<any>;
-
-export interface CreatePullRequestArgs {
-  owner: string;
-  repo: string;
-  title: string;
-  head: string;
-  base: string;
-  body?: string;
-  draft?: boolean;
-}
-
-export interface ListPullRequestsArgs {
-  owner: string;
-  repo: string;
-  state?: "open" | "closed" | "all";
-  head?: string;
-  base?: string;
-  sort?: "created" | "updated" | "popularity" | "long-running";
-  direction?: "asc" | "desc";
-  per_page?: number;
-  page?: number;
-}
-
-export interface GetPullRequestArgs {
-  owner: string;
-  repo: string;
-  pull_number: number;
-}
 
 export const pullRequestHandlers: Record<string, HandlerFunction> = {
   create_pull_request: async (args: unknown) => {
