+++ src/features/search/index.ts
@@ -1,0 +1,3 @@
+export * from "./handler.js";
+export * from "./schema.js";
+export * from "./service.js";
