+++ src/index.ts
@@ -5,7 +5,7 @@ import {
   ListToolsRequestSchema,
 } from "@modelcontextprotocol/sdk/types.js";
 
-import { allHandlers, allToolDefinitions } from "./handlers/index.js";
+import { allHandlers, allToolDefinitions } from "./features/index.js";
 import { serverConfig } from "./config/index.js";
 import { logger } from "./utils/logger.js";
 import { handleError } from "./utils/index.js";
