--- src/services/index.ts
@@ -1,7 +1,0 @@
-export * from "./common.js";
-export * from "./repo.js";
-export * from "./issue.js";
-export * from "./pullRequest.js";
-export * from "./commit.js";
-export * from "./file.js";
-export * from "./search.js";
