+++ src/shared/index.ts
@@ -1,0 +1,2 @@
+export * from "./schema.js";
+export * from "./service.js";
